FY-165/refactor: WaitingRoom 컴포넌트 분리
